+++ src/Controller/Admin/BannersController.php
@@ -19,6 +19,7 @@ use MeCms\Controller\Admin\AppController;
 
 /**
  * Banners controller
+ * @property \MeCms\Model\Table\BannersPositionsTable $Positions
  * @property \MeCms\Model\Table\BannersTable $Banners
  */
 class BannersController extends AppController
@@ -36,7 +37,7 @@ class BannersController extends AppController
         parent::beforeFilter($event);
 
         //Gets positions
-        $positions = $this->Banners->Positions->getList();
+        $positions = $this->Positions->getList();
 
         if ($positions->isEmpty()) {
             $this->Flash->alert(__d('me_cms', 'You must first create a banner position'));
@@ -87,7 +88,8 @@ class BannersController extends AppController
         $this->set('title', I18N_BANNERS);
 
         if ($render) {
-            $this->response = $this->response->withCookie((new Cookie('render-banners', $render))->withNeverExpire());
+            $cookie = (new Cookie('render-banners', $render))->withNeverExpire();
+            $this->setResponse($this->getResponse()->withCookie($cookie));
         }
     }
 
@@ -166,7 +168,7 @@ class BannersController extends AppController
      */
     public function download($id)
     {
-        return $this->response->withFile($this->Banners->get($id)->get('path'), ['download' => true]);
+        return $this->getResponse()->withFile($this->Banners->get($id)->get('path'), ['download' => true]);
     }
 
     /**