+++ src/Controller/PagesController.php
@@ -19,6 +19,7 @@ use MeCms\Utility\StaticPage;
 
 /**
  * Pages controller
+ * @property \MeCms\Model\Table\PagesCategoriesTable $Categories
  * @property \MeCms\Model\Table\PagesTable $Pages
  */
 class PagesController extends AppController
@@ -68,7 +69,7 @@ class PagesController extends AppController
 
         $slug = rtrim($slug, '/');
         $page = $this->Pages->findActiveBySlug($slug)
-            ->contain([$this->Pages->Categories->getAlias() => ['fields' => ['title', 'slug']]])
+            ->contain([$this->Categories->getAlias() => ['fields' => ['title', 'slug']]])
             ->cache('view_' . md5($slug))
             ->firstOrFail();
 
@@ -84,7 +85,7 @@ class PagesController extends AppController
     public function preview($slug)
     {
         $page = $this->Pages->findPendingBySlug($slug)
-            ->contain([$this->Pages->Categories->getAlias() => ['fields' => ['title', 'slug']]])
+            ->contain([$this->Categories->getAlias() => ['fields' => ['title', 'slug']]])
             ->firstOrFail();
 
         $this->set(compact('page'));